feat: add download of VEX
