+++ spog/api/src/advisory.rs
@@ -1,6 +1,8 @@
 use std::collections::HashSet;
 
-use crate::search::{fetch_object, QueryParams, SharedState};
+use crate::search::QueryParams;
+use crate::server::{fetch_object, SharedState};
+use actix_web::web::ServiceConfig;
 use actix_web::{web, HttpResponse, Responder};
 use serde_json::json;
 use spog_model::search::SearchResult;
@@ -8,6 +10,31 @@ use trustification_index::IndexStore;
 use vexination_model::prelude::*;
 
 const MAX_LIMIT: usize = 1_000;
+
+pub(crate) fn configure() -> impl FnOnce(&mut ServiceConfig) {
+    |config: &mut ServiceConfig| {
+        config.service(web::resource("/api/v1/advisory/search").to(search));
+        config.service(web::resource("/api/v1/advisory").to(get));
+    }
+}
+
+#[derive(Debug, serde::Deserialize)]
+pub struct GetParams {
+    pub id: String,
+}
+
+pub async fn get(state: web::Data<SharedState>, params: web::Query<GetParams>) -> impl Responder {
+    let params = params.into_inner();
+    let state = &state.vex;
+    let storage = state.storage.read().await;
+
+    // TODO: Stream
+    if let Some(obj) = fetch_object(&storage, &params.id).await {
+        HttpResponse::Ok().json(obj)
+    } else {
+        HttpResponse::NotFound().finish()
+    }
+}
 
 pub async fn search(state: web::Data<SharedState>, params: web::Query<QueryParams>) -> impl Responder {
     let params = params.into_inner();
