+++ spog/api/src/vulnerability.rs
@@ -1,6 +1,8 @@
 use std::collections::HashMap;
 
-use crate::search::{QueryParams, SharedState};
+use crate::search::QueryParams;
+use crate::server::SharedState;
+use actix_web::web::ServiceConfig;
 use actix_web::{web, HttpResponse, Responder};
 use spog_model::search::{SearchResult, VulnSummary};
 use trustification_index::IndexStore;
@@ -8,9 +10,15 @@ use vexination_model::prelude::*;
 
 const MAX_LIMIT: usize = 1_000;
 
+pub(crate) fn configure() -> impl FnOnce(&mut ServiceConfig) {
+    |config: &mut ServiceConfig| {
+        config.service(web::resource("/api/v1/vulnerability/search").to(search));
+    }
+}
+
 pub async fn search(state: web::Data<SharedState>, params: web::Query<QueryParams>) -> impl Responder {
     let params = params.into_inner();
-    tracing::trace!("Querying VEX using {}", params.q);
+    tracing::info!("Querying VEX using {}", params.q);
     let state = &state.vex;
 
     let index = state.index.read().await;
